change the name of the resulting data frame to "evals"
The name for the resulting data frame is mentioned with 'evals' in the text before the exercise.

Actually it would maybe better to use the name 'evals_fortified' as it leaves the original data set unchanged.

But in that case you would need to change (a) the last two exercises (especially the last one where you are using the new variable 'cls_type') and
(b) to change the name of the resulting data frame in your setup chunk.