convert div hint to standard text and put it before exercise chunk
At the moment [you can't mix markdown and R Markdown](https://github.com/rstudio/learnr/issues/421). Therefore, the solution could not be seen by students.

The solution to progressively reveal hints does not work well as the solution chunks are very long, and the hint window takes on the same size even it is much smaller.

So my suggestion is to write the hint as regular text before the exercise chunk.