change markdown hints to R markdown hints
If you are using a markdown hint, students can't see the solution chunk. Not being able to mix markdown hints with R markdown hints or solution is annoying as markdown hints are generally supposed not to be code but provide additional thoughts students should observe.

As a workaround, I have added general hints as comments in R Markdown chunks. Additionally, I have added coded hints as well. Not ideal, but the only workaround I can see to mix these two kinds of clues.

****

As you can see, I have started again to revise the tutorials in the new structure for the 1st edition of the new book. I hope this is ok, if not please tell me.