Change feedback of wrong answers.
In the original version students get by choosing answers the options 1, 2 or 3 the feedback "Incorrect". But the selection itself is correct, even it is only party correct.

Besides I had to delete "None of the above", because the correct answer "All of the above" IS above this last option.

[Technically one would formulate this questions as a Multiple-Choice question (symbolized with checkboxes instead of radio boxes). But of design errors in the `learnr` packages you cannot proviode correct feedback for all options AND allow to retry the question.
(I plan to report this problem in the question design to RStudio.)