Change the question chunk mc3 (#91)
Using a double negative question is always confusing. The second option is judged as correct in the original question, e.g., this assertion should be **incorrect**. But it is actually correct.

On the other hand, the last option, "None of the above" questions are incorrect, is actually the correct one.