Make it clear that the text is now referring to the `mariokart` data.
I recommend this relatively small adaption so that the example is better understandable. It would be beneficial to revise the lesson: There are several changes between the explanation model (mpg) and the exercise model (`mariokart`). Besides that, some text passages are redundant; the lesson lacks signals whenever the change between the two models occurs. It is difficult for students to follow the text as the corresponding data is not easily accessible but is hidden in another section of this lesson. Also (as a general recommendation), I would suggest - in addition to the glimpse() command - always providing links to the data set explanation page (in this case, a link to https://www.openintro.org/data/index.php?data=mariokart).

My strategy with these PRs is to make the most necessary and easy changes and overhaul the whole lesson later.