add detailed explanations
I tried to explain the (complex) procedure with many new commands similar to the previous tiled visualization.

I deleted the sentence "The x, y, and plane objects were created in code that we aren't showing here due to its complexity." because the code is shown. I tried to explain the code snippets as well.

Again: As I am not a domain expert and not an English native, please correct where necessary!)