Bugfix for 04/04. I have defined the one_poll, one_poll_boot and p_hat objects in the lesson setup so that they can be accessed when needed in the various exercises.
The tutorial claims that "one_poll, ... one_poll_boot,... and p_hat are available in your workspace". However, they are not available in the workspace and it causes many errors. For example, exercises bootstrap_percentile_2 and bootstrap_percentile_3 throw an error because "object 'one_poll_boot' not found".

As a side note: the one_poll_boot object is created at least 7 times in this lesson, likely as partial patches to the object not being defined in the initial lesson setup.