Hash Generation feature added (#159)
Before the "Congratulations" page, there is now a "Submit" page that generates a hash that can be submitted.

As per the discussion in #159, I modified the encoder_ui to require students to enter a name and ID. This name and ID are then encoded in the hash (as an "identifier" type), so that submitted names and IDs could theoretically be used to track tutorial completion.