fix: make hrly_pay accessible in exercise 5 (#271)
* fix: make hrly_pay accessible in exercise 5

The setup chunk for exercise 5 had include = FALSE, which meant the
hrly_pay variable was not accessible in the exercise chunk. Removed
the include = FALSE option to make the variable available.

Fixes #243

* fix: use consistent variable name hrly_pay

Changed instances of hrly_rate in the explanatory text to hrly_pay
to match the variable name used in the code. This improves consistency
and reduces confusion for students.

* fix: make objects accessible in all citizens exercises

Removed include = FALSE from all citizens setup chunks to ensure
objects are accessible in their respective exercises. This includes:
- citizens-setup
- citizens_3-setup
- citizens_4-setup
- citizens_5-setup

Part of fix for #243